--- a/modules/util/config/TrainConfig.py
+++ b/modules/util/config/TrainConfig.py
@@ -108,6 +108,10 @@ class TrainOptimizerConfig(BaseConfig):
     use_grams: False
     use_adopt: False
     use_focus: False
+    use_orthograd: False
+    use_atan2: False
+    use_AdEMAMix: False
+    beta3_ema: float
 
     def __init__(self, data: list[(str, Any, type, bool)]):
         super().__init__(data)
@@ -192,6 +196,10 @@ class TrainOptimizerConfig(BaseConfig):
         data.append(("use_grams", False, bool, False))
         data.append(("use_adopt", False, bool, False))
         data.append(("use_focus", False, bool, False))
+        data.append(("use_orthograd", False, bool, False))
+        data.append(("use_atan2", False, bool, False))
+        data.append(("use_AdEMAMix", False, bool, False))
+        data.append(("beta3_ema", None, float, True))
 
         return TrainOptimizerConfig(data)