--- a/modules/util/config/TrainConfig.py
+++ b/modules/util/config/TrainConfig.py
@@ -113,6 +113,10 @@ class TrainOptimizerConfig(BaseConfig):
     use_AdEMAMix: False
     beta3_ema: float
     variance_reduction: False
+    alpha_grad: float
+    beta1_warmup: int
+    min_beta1: float
+    Simplified_AdEMAMix: False
 
     def __init__(self, data: list[(str, Any, type, bool)]):
         super().__init__(data)
@@ -202,6 +206,10 @@ class TrainOptimizerConfig(BaseConfig):
         data.append(("use_AdEMAMix", False, bool, False))
         data.append(("beta3_ema", None, float, True))
         data.append(("variance_reduction", False, bool, False))
+        data.append(("alpha_grad", None, float, True))
+        data.append(("beta1_warmup", None, int, True))
+        data.append(("min_beta1", None, float, True))
+        data.append(("Simplified_AdEMAMix", False, bool, False))
 
         return TrainOptimizerConfig(data)