--- a/modules/util/config/TrainConfig.py
+++ b/modules/util/config/TrainConfig.py
@@ -112,7 +112,6 @@ class TrainOptimizerConfig(BaseConfig):
     use_atan2: False
     use_AdEMAMix: False
     beta3_ema: float
-    variance_reduction: False
     alpha_grad: float
     beta1_warmup: int
     min_beta1: float
@@ -205,7 +204,6 @@ class TrainOptimizerConfig(BaseConfig):
         data.append(("use_atan2", False, bool, False))
         data.append(("use_AdEMAMix", False, bool, False))
         data.append(("beta3_ema", None, float, True))
-        data.append(("variance_reduction", False, bool, False))
         data.append(("alpha_grad", None, float, True))
         data.append(("beta1_warmup", None, int, True))
         data.append(("min_beta1", None, float, True))