@FurkanGozukara 2025-09-10
hey do you plan this on Kohya as well this sounds intriguing to test