@Koratahiu 2025-09-15
##  Update 1: 

### 🦁 Added `Lion_Adv`

An advanced version of the Lion optimizer.

- **Momentum Compression**: Uses the same 1-bit sign compression cycle as other optimizers. Lion is remarkably tolerant to factorization since it relies on gradient *signs* rather than magnitudes, making 1-bit momentum storage a natural, lossless fit.
- **Variance Reduction**: Implements the variance reduction technique from **[Convergence Analysis of the Lion Optimizer in Centralized and Distributed Settings](https://arxiv.org/abs/2508.12327)**.
  - **Theoretical Benefit**: Accelerates convergence and improves training stability by reducing gradient variance, effectively simulating large-batch training while using smaller batches.
  - **When to use**: Small batch sizes, large and diverse datasets, and long training runs - all scenarios in which gradient variance tends to be high.
  - **Cost**: Requires storing gradients from the previous step → adds extra state. Not ideal for VRAM-constrained fine-tuning, but good enough for LoRA/embedding training or when you have spare VRAM and want faster, stabler Lion training.
  - ⚠️ **Note**: This is a theoretical addition, empirical validation is still yet to be proven, though similar techniques have proven effective on [signSGD](https://arxiv.org/abs/2406.00489).

---

### 🦁📈 Added `Prodigy_Lion_Adv`

Prodigy’s dynamic LR adaptation method (D-Adaptation) isn’t limited to Adam, the original Prodigy paper even tested `Prodigy_SGD`.

So we’ve successfully applied it to Lion:

- Works **flawlessly** with Lion’s sign-based updates.
- Tolerates factorization and 1-bit compression **without issue**, unlike `Prodigy_Adam`, which sometimes breaks under similar conditions (see known issue below).
- Same features as `Lion_Adv`.

---

### Skip `ADAMW_Adv` and `PRODIGY_Adv` First Moment

`ADAMW_Adv` and `PRODIGY_Adv` now skip initialization and update of the first moment (momentum) if beta1 is explicitly set to 0 or left unset (empty).
Fun fact: when we skip the first moment for Adopt it will just reduced to normal Adam, so it doesn't make a sense to do so.

---

### ❗ Known issue with `Prodigy_Adv`

After extensive testing, we found that `Prodigy_adv` is highly sensitive to modifications of gradients or their EMA (first moment), such as Adopt normalization or low-rank factorization. These modifications can cause Prodigy’s internal LR calculations to fail. specifically, the learning rate never increases.
This randomly happens with LoRAs (Finetunes so far so good in my tests), so if you encounter it, just disable factorization or set beta1 to 0 (disabling the first moment).

---

### ❗ Known Issue with Aggressive Learning Rates

Aggressive learning rates can destabilize the factored first moment. This can lead to slower convergence and worse results compared the original optimizer, according to tests by both @dxqb and myself.

To avoid this, it's a good idea to check the Prodigy learning rate when using factorization. This will give you a sense of the correct learning rate threshold.

---