@Koratahiu 2025-09-18
## **📊 Test Area**

---

### 1. `Prodigy_Adv` — SDXL DoRA (R128/A128)

![Prodigy_Adv Chart](https://github.com/user-attachments/assets/72bf6855-6c5d-4e2b-82e7-6054a6d2f3a5)

- **🟢 Unfactored**  
- **🔷 Factored**

---

### 2. `Lion_Prodigy_adv` — SDXL DoRA (R128/A128)

![Lion_Prodigy_adv Chart](https://github.com/user-attachments/assets/4c632ce0-c861-4fe3-9138-c604c8691e51)

- **🟣 Unfactored**  
- **🔷 Factored**

---

### 3. `Prodigy_adv` with `Simplified_AdEMAMix` — SDXL DoRA (R64/A1)

![Prodigy_adv + Simplified_AdEMAMix Chart](https://github.com/user-attachments/assets/e9e6ee99-3a74-4e68-b19f-b9c61cc89b11)

- **🔷 With Simplified_AdEMAMix**  
- **🟢 Without**

---

### 4. `Prodigy_adv` + `Simplified_AdEMAMix` — Factored vs Unfactored (R64/A1 SDXL DoRA)

![Prodigy_adv Factored Comparison](https://github.com/user-attachments/assets/7855bf99-f552-4fe2-8c63-a123434dea67)

- **🟠 Unfactored**  
- **🟢 Factored**

---